Recording logic complete and bug free

+++ src/audio/Play.py
@@ -1,5 +1,5 @@
 import numpy as np
-from numpy import int16
+from numpy import float32
 from numpy.typing import NDArray
 import sounddevice as sd
 
@@ -8,7 +8,7 @@ class Player():
     def __init__(self, samplerate: int = 44100) -> None:
         self.samplerate = samplerate
 
-    def play(self, wave: NDArray[int16], attack: float = 0.5, decay: float = 0.5) -> None:
+    def play(self, wave: NDArray[float32], attack: float = 0.5, decay: float = 0.5) -> None:
         wave_1d = wave.flatten()
         total = len(wave_1d)
 
@@ -19,5 +19,5 @@ class Player():
         env[:n_attack] = np.linspace(0, 1, n_attack)
         env[-n_decay:] = np.linspace(1, 0, n_decay)
 
-        sd.play((wave_1d * env).astype(np.int16), self.samplerate)
+        sd.play((wave_1d * env).astype(np.float32), self.samplerate)
         sd.wait()
